ctp.py: fixed misc issues found while testing
Made the string representation of the CtpSyntaxError return the
message string, so it becomes easier to match for with
assertRaisesRegex.

Removed extra ' from a message

Made the error message for not providing a value for all pins
in a CHECK command more general. This is for the case where the
user forgets more than one pin, where in the old case the script
would only report one of the pins.