start distributing on PyPi
This is to make it easier to get the project.
To this end a setup.py file was created to use the
setuptools module. This file is not yet final as docopt
have not been made a requirement (I know).

Furthermore, since setuptools makes a new dir with some
files .gitignore has been told to ignore that. At least
for now.